refactor: created components for some ui elements, extracted formatTile func from hook, updated ui
